Change to long

+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
 
         SalesManager salesManager = new SalesManager();
-        int[] saleStat = new int[]{1, 2, 3, 4};
+        long[] saleStat = new long[]{1, 2, 3, 4};
         salesManager.sales = saleStat;
         System.out.println(salesManager.max());
     }
